Spring Data JPA Preloading data with CommandLineRunner
