+++ src/main/java/tacocloud/HomeController.java
@@ -1,0 +1,12 @@
+package tacocloud;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+    @GetMapping("/")
+    public String home() {
+        return "home";
+    }
+}
