+++ src/main/java/tacocloud/IngredientRef.java
@@ -1,0 +1,8 @@
+package tacos;
+
+import lombok.Data;
+
+@Data
+public class IngredientRef {
+    private String ingredient;// final
+}
