+++ src/main/java/tacocloud/repository/OrderRepository.java
@@ -1,0 +1,7 @@
+package tacocloud.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import tacocloud.model.TacoOrder;
+
+public interface OrderRepository extends CrudRepository<TacoOrder, Long> {
+}
