+++ src/main/java/tacocloud/repository/TacoRepository.java
@@ -1,0 +1,7 @@
+package tacocloud.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import tacocloud.model.Taco;
+
+public interface TacoRepository extends CrudRepository<Taco, Long> {
+}
