+++ src/main/java/tacocloud/web/IngredientByIdConverter.java
@@ -1,0 +1,21 @@
+package tacocloud.web;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.convert.converter.Converter;
+import org.springframework.stereotype.Component;
+import tacocloud.model.Ingredient;
+import tacocloud.repository.IngredientRepository;
+
+@Component// Сделать видимым для spring
+public class IngredientByIdConverter implements Converter<String, Ingredient> {
+
+    private IngredientRepository ingredientRepo;
+    @Autowired
+    public IngredientByIdConverter(IngredientRepository ingredientRepo) {
+        this.ingredientRepo = ingredientRepo;
+    }
+    @Override
+    public Ingredient convert(String id) {
+        return ingredientRepo.findById(id).orElse(null);
+    }
+}
