--- src/main/java/tacos/IngredientRef.java
@@ -1,8 +1,0 @@
-package tacos;
-
-import lombok.Data;
-
-@Data
-public class IngredientRef {
-    private String ingredient;// final
-}
