--- src/main/java/tacos/data/JdbcIngredientRepository.java
@@ -1,60 +1,0 @@
-package tacos.data;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Repository;
-import tacos.Ingredient;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.List;
-import java.util.Optional;
-
-//not used
-//@Repository //Сделать видимым для spring и создать bean
-public class JdbcIngredientRepository {//implements IngredientRepository {
-
-    private JdbcTemplate jdbc;
-
-    // Spring внедряет JdbcTemplate через autowired
-    @Autowired
-    public JdbcIngredientRepository(JdbcTemplate jdbc) {
-        this.jdbc = jdbc;
-    }
-
-//    @Override
-    public List<Ingredient> findAll() {
-        return jdbc.query("select id, name, type from Ingredient", this::mapRowToIngredient);
-    }
-
-//    @Override
-    public Optional<Ingredient> findOne(String id) {
-//        return jdbc.queryForObject("select id, name, type from Ingredient where id=?", this::mapRowToIngredient, id);
-        List<Ingredient> results = jdbc.query(
-                "select id, name, type from Ingredient where id=?",
-                this::mapRowToIngredient,
-                id);
-        return results.size() == 0 ?
-                Optional.empty() :
-                Optional.of(results.get(0));
-    }
-
-    //Two ways to save data with JdbcTemplate
-    //with update() method
-    public Ingredient save(Ingredient ingredient) {
-        jdbc.update(
-                "insert into Ingredient (id, name, type) values (?, ?, ?)",
-                ingredient.getId(),
-                ingredient.getName(),
-                ingredient.getType().toString());
-        return ingredient;
-    }
-    //with SimpleJdbcInsert wrapper class
-
-    private Ingredient mapRowToIngredient(ResultSet rs, int rowNum) throws SQLException {
-        return new Ingredient(
-                rs.getString("id"),
-                rs.getString("name"),
-                Ingredient.Type.valueOf(rs.getString("type")));
-    }
-}
