--- src/main/java/tacos/data/JdbcOrderRepository.java
@@ -1,139 +1,0 @@
-package tacos.data;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcOperations;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import tacos.TacoOrder;
-import tacos.Taco;
-
-import java.util.*;
-
-//not used
-//@Repository
-public class JdbcOrderRepository {//implements OrderRepository{
-
-    private JdbcOperations jdbcOperations;
-    private JdbcTemplate jdbc;
-    private SimpleJdbcInsert orderInserter;
-    private SimpleJdbcInsert orderTacoInserter;
-    private ObjectMapper objectMapper;
-// 1 способ
-//    public JdbcOrderRepository(JdbcOperations jdbcOperations) {
-//        this.jdbcOperations = jdbcOperations;
-//    }
-
-//  2 способ с SimpleJdbcInsert
-    @Autowired
-    public JdbcOrderRepository(JdbcTemplate jdbc) {
-        this.orderInserter = new SimpleJdbcInsert(jdbc)
-                .withTableName("Taco_Order")
-                .usingGeneratedKeyColumns("id");
-        this.orderTacoInserter = new SimpleJdbcInsert(jdbc)
-                .withTableName("Taco_Order_Tacos");
-        this.objectMapper = new ObjectMapper();
-    }
-
-// 1 способ с методом update()
-//    @Override
-//    @Transactional
-//    public Order save(Order order) {
-//        // Запрос
-//        PreparedStatementCreatorFactory pscf =
-//                new PreparedStatementCreatorFactory(
-//                        "insert into Taco_Order "
-//                                + "(delivery_name, delivery_street, delivery_city, "
-//                                + "delivery_state, delivery_zip, cc_number, "
-//                                + "cc_expiration, cc_cvv, placed_at) "
-//                                + "values (?,?,?,?,?,?,?,?,?)",
-//                        Types.VARCHAR, Types.VARCHAR, Types.VARCHAR,
-//                        Types.VARCHAR, Types.VARCHAR, Types.VARCHAR,
-//                        Types.VARCHAR, Types.VARCHAR, Types.TIMESTAMP
-//                );
-//        pscf.setReturnGeneratedKeys(true);
-//        order.setCreatedAt(new Date());
-//        // Передаем переменные в запрос
-//        PreparedStatementCreator psc =
-//                pscf.newPreparedStatementCreator(
-//                        Arrays.asList(
-//                                order.getName(),
-//                                order.getStreet(),
-//                                order.getCity(),
-//                                order.getState(),
-//                                order.getZip(),
-//                                order.getCcNumber(),
-//                                order.getCcExpiration(),
-//                                order.getCcCVV(),
-//                                order.getCreatedAt()));
-//        GeneratedKeyHolder keyHolder = new GeneratedKeyHolder();
-//        jdbcOperations.update(psc, keyHolder);
-//        long orderId = keyHolder.getKey().longValue();
-//        order.setId(orderId);
-//        List<Taco> tacos = order.getTacos();
-//        int i=0;
-//        for (Taco taco : tacos) {
-//            saveTaco(orderId, i++, taco);
-//        }
-//        return order;
-//    }
-//    private long saveTaco(Long orderId, int orderKey, Taco taco) {
-//        taco.setCreatedAt(new Date());
-//        PreparedStatementCreatorFactory pscf =
-//                new PreparedStatementCreatorFactory(
-//                        "insert into Taco "
-//                                + "(name, created_at, taco_order, taco_order_key) "
-//                                + "values (?, ?, ?, ?)",
-//                        Types.VARCHAR, Types.TIMESTAMP, Type.LONG, Type.LONG
-//                );
-//        pscf.setReturnGeneratedKeys(true);
-//        PreparedStatementCreator psc =
-//                pscf.newPreparedStatementCreator(
-//                        Arrays.asList(
-//                                taco.getName(),
-//                                taco.getCreatedAt(),
-//                                orderId,
-//                                orderKey));
-//        GeneratedKeyHolder keyHolder = new GeneratedKeyHolder();
-//        jdbcOperations.update(psc, keyHolder);
-//        long tacoId = keyHolder.getKey().longValue();
-//        taco.setId(tacoId);
-//        saveIngredientRefs(tacoId, taco.getIngredients());
-//        return tacoId;
-//    }
-//    private void saveIngredientRefs(long tacoId, List<Ingredient> ingredientRefs) {
-//        int key = 0;
-//        for (Ingredient ingredientRef : ingredientRefs) {
-//            jdbcOperations.update(
-//                    "insert into Ingredient_Ref (ingredient, taco, taco_key) "
-//                            + "values (?, ?, ?)",
-//                    ingredientRef, tacoId, key++);
-//        }
-//    }
-//  2 способ с SimpleJdbcInsert
-//    @Override
-    public TacoOrder save(TacoOrder tacoOrder) {
-        tacoOrder.setCreatedAt(new Date());
-        long orderId = saveOrderDetails(tacoOrder);
-        tacoOrder.setId(orderId);
-        List<Taco> tacos = tacoOrder.getTacos();
-        for (Taco taco : tacos) {
-            saveTacoToOrder(taco, orderId);
-        }
-        return tacoOrder;
-    }
-    private long saveOrderDetails(TacoOrder tacoOrder) {// 4009736722384759
-        @SuppressWarnings("unchecked")
-        Map<String, Object> values = objectMapper.convertValue(tacoOrder, Map.class);
-        values.put("createdAt", tacoOrder.getCreatedAt());
-        long orderId =orderInserter.executeAndReturnKey(values).longValue();
-        return orderId;
-    }
-    private void saveTacoToOrder(Taco taco, long orderId) {
-        Map<String, Object> values = new HashMap<>();
-        values.put("tacoOrder", orderId);
-        values.put("taco", taco.getId());
-        orderTacoInserter.execute(values);
-    }
-
-}
