--- src/main/java/tacos/data/JdbcTacoRepository.java
@@ -1,66 +1,0 @@
-package tacos.data;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStatementCreator;
-import org.springframework.jdbc.core.PreparedStatementCreatorFactory;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-import tacos.Ingredient;
-import tacos.Taco;
-
-import java.sql.Timestamp;
-import java.sql.Types;
-import java.util.Arrays;
-import java.util.Date;
-
-//not used
-@Repository
-public class JdbcTacoRepository implements TacoRepository{
-    private JdbcTemplate jdbc;
-    public JdbcTacoRepository(JdbcTemplate jdbc) {
-        this.jdbc = jdbc;
-    }
-//=====================saving with update() method
-    @Override
-    public Taco save(Taco taco) {
-        long tacoId = saveTacoInfo(taco);
-        taco.setId(tacoId);
-        for (Ingredient ingredient : taco.getIngredients()) {
-            saveIngredientToTaco(ingredient, tacoId);
-        }
-        return taco;
-    }
-    // NULL not allowed for column "TACO_ORDER"
-    private long saveTacoInfo(Taco taco) {
-        taco.setCreatedAt(new Date());
-//        PreparedStatementCreator psc =
-//                new PreparedStatementCreatorFactory(
-//                        "insert into Taco (name, created_at) values (?, ?)",
-//                        Types.VARCHAR, Types.TIMESTAMP
-//                ).newPreparedStatementCreator(
-//                        Arrays.asList(
-//                                taco.getName(),
-//                                new Timestamp(taco.getCreatedAt().getTime())));
-        PreparedStatementCreatorFactory pscf = new PreparedStatementCreatorFactory(
-                "insert into Taco (name, created_at) values (?, ?)",
-                Types.VARCHAR, Types.TIMESTAMP);
-
-// By default, returnGeneratedKeys = false so change it to true
-        pscf.setReturnGeneratedKeys(true);
-
-        PreparedStatementCreator psc =
-                pscf.newPreparedStatementCreator(
-                        Arrays.asList(
-                                taco.getName(),
-                                new Timestamp(taco.getCreatedAt().getTime())));
-        // Cгенерированный идентификатор taco
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbc.update(psc, keyHolder);
-        return keyHolder.getKey().longValue();
-    }
-    private void saveIngredientToTaco(Ingredient ingredient, long tacoId) {
-        jdbc.update("insert into Ingredient (id, name, type) values (?, ?, ?)", tacoId, ingredient.getName(), ingredient.getType().toString());
-    }
-//=================================================
-}
