--- src/main/java/tacos/data/TacoRepository.java
@@ -1,8 +1,0 @@
-package tacos.data;
-
-import org.springframework.data.repository.CrudRepository;
-import tacos.Taco;
-
-public interface TacoRepository {//extends CrudRepository<Taco, Long> {
-    Taco save(Taco design);
-}
