--- src/main/java/tacos/web/IngredientByIdConverter.java
@@ -1,24 +1,0 @@
-package tacos.web;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.stereotype.Component;
-import tacos.Ingredient;
-import tacos.data.IngredientRepository;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Component// Сделать видимым для spring
-public class IngredientByIdConverter implements Converter<String, Ingredient> {
-
-    private IngredientRepository ingredientRepo;
-    @Autowired
-    public IngredientByIdConverter(IngredientRepository ingredientRepo) {
-        this.ingredientRepo = ingredientRepo;
-    }
-    @Override
-    public Ingredient convert(String id) {
-        return ingredientRepo.findById(id).orElse(null);
-    }
-}
