+++ src/main/resources/schema.sql
@@ -14,20 +14,19 @@ create table if not exists Taco (
     id identity,
     name varchar(50) not null,
     taco_order bigint,
-    taco_order_key bigint,
+--     taco_order_key bigint,
     created_at timestamp not null
 );
 create table if not exists Ingredient_Ref (
-    ingredient varchar(4) not null,
     taco bigint not null,
-    taco_key bigint not null
+    ingredient varchar(4) not null
+--     taco_key bigint not null
 );
 create table if not exists Ingredient (
     id varchar(4) not null,
     name varchar(25) not null,
     type varchar(10)
 );
-alter table Taco
-    add foreign key (taco_order) references Taco_Order(id);
-alter table Ingredient_Ref
-    add foreign key (ingredient) references Ingredient(id);
+alter table Taco add foreign key (taco_order) references Taco_Order(id);
+alter table Ingredient_Ref add foreign key (taco) references Taco(id);
+alter table Ingredient_Ref add foreign key (ingredient) references Ingredient(id);
