+++ src/main/resources/templates/orderForm.html
@@ -11,7 +11,7 @@
         <form method="POST" th:action="@{/orders}" th:object="${tacoOrder}">
             <h1>Order your taco creations!</h1>
             <img th:src="@{/images/TacoCloud.jpg}"/><br/>
-            <h3>Your tacos in this tacoOrder:</h3>
+            <h3>Your tacocloud in this tacoOrder:</h3>
             <a th:href="@{/design}" id="another">Design another taco</a><br/>
             <ul>
                 <li th:each="taco : ${tacoOrder.tacos}">
