+++ src/test/java/tacocloud/TacoCloudApplicationTests.java
@@ -1,0 +1,14 @@
+package tacocloud;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class TacoCloudApplicationTests {
+    @Test
+    public void contextLoads() {
+    }
+}
